adding the first bit of encryption work.
This commit does not yet include any file encryption. merely the generation of public and private keys at user verification time.
the pub key is b64'd and stored in the db.
the priv key is AES encrypted using the user's passphrase, then b64'd, then stored in the db.