Moving key management out of the user definition.
This migration allows for a user to manage/export/delete keypairs.
It's a first step towards implementing issue #4 but still is missing a few things.
The biggest piece left to solve it is simply moving the notion of sharing
out of the User model and into the key. As a temporary step, right now
sharing happens across all keys of a given user, whereas in the future
solution sharing will be strictly to specified keys.